#219 feat: storybook MenuList 작성
